Improve sdl audio implementation
Also renamed and moved code from utils
