+++ sdl/src/audio/mod.rs
@@ -1,7 +1,7 @@
 mod multi_device_audio;
 mod wav_file_audio_device;
-mod sdl_pull_audio_device;
+mod sdl_audio_device;
 
 pub use multi_device_audio::MultiAudioDevice;
 pub use wav_file_audio_device::WavfileAudioDevice;
-pub use sdl_pull_audio_device::SdlPullAudioDevice;
+pub use sdl_audio_device::SdlAudioDevice;
